AI/envs/pyproject.toml
[project]
name = "env_2048"
version = "0.1.0"
dependencies = ["gymnasium", "pygame-ce"]
